parser/requirements.txt
geopandas
jsonschema
ezdxf
opencv-python
python-dotenv
ibm-watsonx-ai
